Split AccountUpdater into implementation and interface
