+++ src/main/java/org/javacord/AccountUpdater.java
@@ -1,48 +1,17 @@
 package org.javacord;
 
-import com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.javacord.entity.Icon;
-import org.javacord.util.FileContainer;
-import org.javacord.util.rest.RestEndpoint;
-import org.javacord.util.rest.RestMethod;
-import org.javacord.util.rest.RestRequest;
 
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.InputStream;
 import java.net.URL;
-import java.util.Base64;
 import java.util.concurrent.CompletableFuture;
 
 /**
- * This class can be used to update the connected account (e.g. username or avatar).
+ * This interface can be used to update the connected account (e.g. username or avatar).
  */
-public class AccountUpdater {
-
-    /**
-     * The discord api instance.
-     */
-    private final DiscordApi api;
-
-    /**
-     * The username to update.
-     */
-    private String username = null;
-
-    /**
-     * The avatar to update.
-     */
-    private FileContainer avatar = null;
-
-    /**
-     * Creates a new account updater.
-     *
-     * @param api The discord api instance.
-     */
-    public AccountUpdater(DiscordApi api) {
-        this.api = api;
-    }
+public interface AccountUpdater {
 
     /**
      * Queues the username of the connected account to get updated.
@@ -50,10 +19,7 @@ public class AccountUpdater {
      * @param username The username to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setUsername(String username) {
-        this.username = username;
-        return this;
-    }
+    AccountUpdater setUsername(String username);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -62,10 +28,7 @@ public class AccountUpdater {
      * @param avatar The avatar to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(BufferedImage avatar) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar, "png");
-        return this;
-    }
+    AccountUpdater setAvatar(BufferedImage avatar);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -74,10 +37,7 @@ public class AccountUpdater {
      * @param fileType The type of the avatar, e.g. "png" or "jpg".
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(BufferedImage avatar, String fileType) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar, fileType);
-        return this;
-    }
+    AccountUpdater setAvatar(BufferedImage avatar, String fileType);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -85,10 +45,7 @@ public class AccountUpdater {
      * @param avatar The avatar to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(File avatar) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar);
-        return this;
-    }
+    AccountUpdater setAvatar(File avatar);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -96,10 +53,7 @@ public class AccountUpdater {
      * @param avatar The avatar to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(Icon avatar) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar);
-        return this;
-    }
+    AccountUpdater setAvatar(Icon avatar);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -107,10 +61,7 @@ public class AccountUpdater {
      * @param avatar The avatar to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(URL avatar) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar);
-        return this;
-    }
+    AccountUpdater setAvatar(URL avatar);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -119,10 +70,7 @@ public class AccountUpdater {
      * @param avatar The avatar to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(byte[] avatar) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar, "png");
-        return this;
-    }
+    AccountUpdater setAvatar(byte[] avatar);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -131,10 +79,7 @@ public class AccountUpdater {
      * @param fileType The type of the avatar, e.g. "png" or "jpg".
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(byte[] avatar, String fileType) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar, fileType);
-        return this;
-    }
+    AccountUpdater setAvatar(byte[] avatar, String fileType);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -143,10 +88,7 @@ public class AccountUpdater {
      * @param avatar The avatar to set.
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(InputStream avatar) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar, "png");
-        return this;
-    }
+    AccountUpdater setAvatar(InputStream avatar);
 
     /**
      * Queues the avatar of the connected account to get updated.
@@ -155,44 +97,13 @@ public class AccountUpdater {
      * @param fileType The type of the avatar, e.g. "png" or "jpg".
      * @return The current instance in order to chain call methods.
      */
-    public AccountUpdater setAvatar(InputStream avatar, String fileType) {
-        this.avatar = (avatar == null) ? null : new FileContainer(avatar, fileType);
-        return this;
-    }
+    AccountUpdater setAvatar(InputStream avatar, String fileType);
 
     /**
      * Performs the queued updates.
      *
      * @return A future to check if the update was successful.
      */
-    public CompletableFuture<Void> update() {
-        boolean patchAccount = false;
-        ObjectNode body = JsonNodeFactory.instance.objectNode();
-        if (username != null) {
-            body.put("username", username);
-            patchAccount = true;
-        }
-        if (avatar != null) {
-            patchAccount = true;
-        }
-        if (patchAccount) {
-            if (avatar != null) {
-                return avatar.asByteArray(api).thenAccept(bytes -> {
-                    String base64Avatar = "data:image/" + avatar.getFileType() + ";base64," +
-                            Base64.getEncoder().encodeToString(bytes);
-                    body.put("avatar", base64Avatar);
-                }).thenCompose(aVoid -> new RestRequest<Void>(api, RestMethod.PATCH, RestEndpoint.CURRENT_USER)
-                        .setRatelimitRetries(0)
-                        .setBody(body)
-                        .execute(result -> null));
-            }
-            return new RestRequest<Void>(api, RestMethod.PATCH, RestEndpoint.CURRENT_USER)
-                    .setRatelimitRetries(0)
-                    .setBody(body)
-                    .execute(result -> null);
-        } else {
-            return CompletableFuture.completedFuture(null);
-        }
-    }
+    CompletableFuture<Void> update();
 
 }
